gateway/requirements.txt
fastapi>=0.135,<1
PyJWT>=2.12,<3
uvicorn>=0.38,<1
